Extracted result synchronization data provider and formatter
