+++ config/default/SyncResultDataFormatter.conf.php
@@ -1,0 +1,5 @@
+<?php
+
+use oat\taoSync\model\SyncResultDataFormatter;
+
+return new SyncResultDataFormatter();
